# Optimised Hybrid Search CPU