--- a/cpp/open3d/core/nns/NanoFlannIndex.cpp
+++ b/cpp/open3d/core/nns/NanoFlannIndex.cpp
@@ -26,6 +26,7 @@
 
 #include "open3d/core/nns/NanoFlannIndex.h"
 
+#include <tbb/blocked_range2d.h>
 #include <tbb/parallel_for.h>
 
 #include <nanoflann.hpp>
@@ -233,10 +234,7 @@ std::tuple<Tensor, Tensor, Tensor> NanoFlannIndex::SearchRadius(
 
 std::pair<Tensor, Tensor> NanoFlannIndex::SearchHybrid(
         const Tensor &query_points, double radius, int max_knn) const {
-    // Check dtype.
     query_points.AssertDtype(GetDtype());
-
-    // Check shapes.
     query_points.AssertShapeCompatible({utility::nullopt, GetDimension()});
 
     if (max_knn <= 0) {
@@ -251,57 +249,55 @@ std::pair<Tensor, Tensor> NanoFlannIndex::SearchHybrid(
     }
 
     double radius_squared = radius * radius;
-
-    Tensor indices;
-    Tensor distances;
-    std::tie(indices, distances) = SearchKnn(query_points, max_knn);
-
+    int64_t num_query_points = query_points.GetShape()[0];
+    Tensor indices, distances;
     Dtype dtype = GetDtype();
+
     DISPATCH_FLOAT32_FLOAT64_DTYPE(dtype, [&]() {
-        Tensor invalid = distances.Gt(radius_squared);
-        Tensor invalid_indices =
-                Tensor(std::vector<int64_t>({-1}), {1}, indices.GetDtype(),
-                       indices.GetDevice());
-        Tensor invalid_distances =
-                Tensor(std::vector<scalar_t>({-1}), {1}, distances.GetDtype(),
-                       distances.GetDevice());
+        indices = Tensor::Empty({num_query_points, max_knn}, Dtype::Int64);
+        auto indices_ptr = indices.GetDataPtr<int64_t>();
+        distances = Tensor::Empty({num_query_points, max_knn}, dtype);
+        auto distances_ptr = distances.GetDataPtr<scalar_t>();
 
-        indices.SetItem(TensorKey::IndexTensor(invalid), invalid_indices);
-        distances.SetItem(TensorKey::IndexTensor(invalid), invalid_distances);
+        auto holder = static_cast<NanoFlannIndexHolder<L2, scalar_t> *>(
+                holder_.get());
+
+        nanoflann::SearchParams params;
+
+        // Parallel search.
+        tbb::parallel_for(
+                tbb::blocked_range<size_t>(0, num_query_points),
+                [&](const tbb::blocked_range<size_t> &r) {
+                    std::vector<std::pair<int64_t, scalar_t>> ret_matches;
+                    for (size_t workload_idx = r.begin();
+                         workload_idx != r.end(); ++workload_idx) {
+                        int64_t result_idx = workload_idx * max_knn;
+
+                        size_t num_results = holder->index_->radiusSearch(
+                                static_cast<scalar_t *>(
+                                        query_points[workload_idx]
+                                                .GetDataPtr()),
+                                radius_squared, ret_matches, params);
+                        ret_matches.resize(num_results);
+
+                        int neighbour_idx = 0;
+                        for (auto it = ret_matches.begin();
+                             it < ret_matches.end() && neighbour_idx < max_knn;
+                             it++, neighbour_idx++) {
+                            indices_ptr[result_idx + neighbour_idx] = it->first;
+                            distances_ptr[result_idx + neighbour_idx] =
+                                    it->second;
+                        }
+
+                        while (neighbour_idx < max_knn) {
+                            indices_ptr[result_idx + neighbour_idx] = -1;
+                            distances_ptr[result_idx + neighbour_idx] = -1;
+                            neighbour_idx += 1;
+                        }
+                    }
+                });
     });
-
     return std::make_pair(indices, distances);
-}
-
-std::tuple<Tensor, Tensor, Tensor> NanoFlannIndex::SearchHybrid1NN(
-        const Tensor &query_points, double radius) const {
-    Device device = GetDevice();
-    query_points.AssertDtype(GetDtype());
-    query_points.AssertShapeCompatible({utility::nullopt, GetDimension()});
-
-    if (radius <= 0) {
-        utility::LogError(
-                "[NanoFlannIndex::SearchHybrid] radius should be larger than "
-                "0.");
-    }
-
-    double radius_squared = radius * radius;
-
-    Tensor indices, distances, source_indices, target_indices;
-    std::tie(indices, distances) = SearchKnn(query_points, 1);
-
-    Tensor valid = distances.Le(radius_squared).Reshape({-1});
-    // correpondence_set : (i, corres[i]).
-    // source[i] and target[corres[i]] is a correspondence.
-    source_indices = core::Tensor::Arange(0, query_points.GetShape()[0], 1,
-                                          core::Dtype::Int64, device)
-                             .IndexGet({valid});
-    // Only take valid indices.
-    target_indices = indices.IndexGet({valid}).Reshape({-1});
-    // Only take valid distances.
-    distances = distances.IndexGet({valid});
-
-    return std::make_tuple(source_indices, target_indices, distances);
 }
 
 }  // namespace nns