Performance on PR 3122 [NNS cpp benchmark]:
```
------------------------------------------------------------------------------------------
Benchmark                                                Time             CPU   Iterations
------------------------------------------------------------------------------------------
BM_TestNNS/CPU/1000/1                                 2701 ms         2688 ms            1
BM_TestNNS/CPU/1000/5                                 4308 ms         4308 ms            1
```
Performance of this PR on same benchmark example:
```
------------------------------------------------------------------------------------------
Benchmark                                                Time             CPU   Iterations
------------------------------------------------------------------------------------------
BM_TestNNS/CPU/1000/1                                  136 ms          136 ms            5
BM_TestNNS/CPU/1000/5                                  140 ms          140 ms            5